llm-judge-pki/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
datasets>=2.19.0
tenacity>=8.2.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.13.0
